requirements.txt
requests==2.31.0
sseclient-py==1.8.0
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
